Feat/logic

+++ internal/server/handler/round/handler.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"errors"
+	"time"
 
 	"fundlevel/internal/entities/round"
 	"fundlevel/internal/server/handler/shared"
@@ -38,11 +39,49 @@ type CreateRoundInput struct {
 }
 
 func (i *CreateRoundInput) Resolve(ctx huma.Context) []error {
-	//TODO: implement db checks
+
+	if i.Body.BeginsAt.Before(time.Now().AddDate(0, 0, 1)) {
+		return []error{&huma.ErrorDetail{
+			Message:  "begins at must be in the future",
+			Location: "round.beginsAt",
+			Value:    i.Body.BeginsAt,
+		}}
+	}
+
+	if i.Body.EndsAt.Before(i.Body.BeginsAt) {
+		return []error{&huma.ErrorDetail{
+			Message:  "ends at must be after begins at",
+			Location: "round.endsAt",
+			Value:    i.Body.EndsAt,
+		}}
+	}
+
 	return nil
 }
 
 func (h *httpHandler) create(ctx context.Context, input *CreateRoundInput) (*shared.SingleRoundResponse, error) {
+	venture, err := h.service.VentureService.GetById(ctx, input.Body.VentureID)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			return nil, huma.Error404NotFound("Venture ID not found")
+		}
+
+		h.logger.Error("failed to fetch venture", zap.Error(err))
+		return nil, huma.Error500InternalServerError("An error occurred while fetching the venture")
+	}
+
+	if account := shared.GetAuthenticatedAccount(ctx); account.ID != venture.Business.OwnerAccountID {
+		h.logger.Error("business owner account id does not match authenticated account id",
+			zap.Any("business owner account id", venture.Business.OwnerAccountID),
+			zap.Any("authenticated account id", account.ID))
+
+		return nil, huma.Error403Forbidden("Cannot create round for a business you do not own")
+	}
+
+	if venture.IsHidden {
+		return nil, huma.Error400BadRequest("Cannot create round for a hidden venture")
+	}
+
 	round, err := h.service.RoundService.Create(ctx, input.Body)
 	if err != nil {
 		h.logger.Error("failed to create round", zap.Error(err))
@@ -57,7 +96,7 @@ func (h *httpHandler) create(ctx context.Context, input *CreateRoundInput) (*sha
 }
 
 func (h *httpHandler) delete(ctx context.Context, input *shared.PathIDParam) (*shared.MessageResponse, error) {
-	_, err := h.service.RoundService.GetById(ctx, input.ID)
+	round, err := h.service.RoundService.GetById(ctx, input.ID)
 	if err != nil {
 		switch {
 		case errors.Is(err, sql.ErrNoRows):
@@ -66,6 +105,14 @@ func (h *httpHandler) delete(ctx context.Context, input *shared.PathIDParam) (*s
 			h.logger.Error("failed to fetch round", zap.Error(err))
 			return nil, huma.Error500InternalServerError("An error occurred while fetching the round")
 		}
+	}
+
+	if account := shared.GetAuthenticatedAccount(ctx); account.ID != round.Venture.Business.OwnerAccountID {
+		h.logger.Error("business owner account id does not match authenticated account id",
+			zap.Any("business owner account id", round.Venture.Business.OwnerAccountID),
+			zap.Any("authenticated account id", account.ID))
+
+		return nil, huma.Error403Forbidden("Cannot delete round for a business you do not own")
 	}
 
 	err = h.service.RoundService.Delete(ctx, input.ID)
@@ -80,10 +127,29 @@ func (h *httpHandler) delete(ctx context.Context, input *shared.PathIDParam) (*s
 	return resp, nil
 }
 
-func (h *httpHandler) getInvestmentsByCursor(ctx context.Context, input *shared.GetCursorPaginatedByParentPathIDInput) (*shared.GetCursorPaginatedRoundInvestmentsOutput, error) {
+func (h *httpHandler) getInvestmentsByCursor(ctx context.Context, input *shared.GetInvestmentsByParentAndCursorInput) (*shared.GetCursorPaginatedRoundInvestmentsOutput, error) {
+	round, err := h.service.RoundService.GetById(ctx, input.ID)
+	if err != nil {
+		switch {
+		case errors.Is(err, sql.ErrNoRows):
+			return nil, huma.Error404NotFound("round not found")
+		default:
+			h.logger.Error("failed to fetch round", zap.Error(err))
+			return nil, huma.Error500InternalServerError("An error occurred while fetching the round")
+		}
+	}
+
+	if account := shared.GetAuthenticatedAccount(ctx); account.ID != round.Venture.Business.OwnerAccountID {
+		h.logger.Error("business owner account id does not match authenticated account id",
+			zap.Any("business owner account id", round.Venture.Business.OwnerAccountID),
+			zap.Any("authenticated account id", account.ID))
+
+		return nil, huma.Error403Forbidden("Cannot get investments for a round for a business you do not own")
+	}
+
 	limit := input.Limit + 1
 
-	investments, err := h.service.RoundService.GetInvestmentsByCursor(ctx, input.ID, limit, input.Cursor)
+	investments, err := h.service.RoundService.GetInvestmentsByCursor(ctx, input.ID, limit, input.Cursor, input.InvestmentFilter)
 
 	if err != nil {
 		switch {
@@ -108,9 +174,27 @@ func (h *httpHandler) getInvestmentsByCursor(ctx context.Context, input *shared.
 	return resp, nil
 }
 
-func (h *httpHandler) getInvestmentsByPage(ctx context.Context, input *shared.GetOffsetPaginatedByParentPathIDInput) (*shared.GetOffsetPaginatedRoundInvestmentsOutput, error) {
+func (h *httpHandler) getInvestmentsByPage(ctx context.Context, input *shared.GetInvestmentsByParentAndPageInput) (*shared.GetOffsetPaginatedRoundInvestmentsOutput, error) {
+	round, err := h.service.RoundService.GetById(ctx, input.ID)
+	if err != nil {
+		switch {
+		case errors.Is(err, sql.ErrNoRows):
+			return nil, huma.Error404NotFound("round not found")
+		default:
+			h.logger.Error("failed to fetch round", zap.Error(err))
+			return nil, huma.Error500InternalServerError("An error occurred while fetching the round")
+		}
+	}
 
-	investments, err := h.service.RoundService.GetInvestmentsByPage(ctx, input.ID, input.PageSize, input.Page)
+	if account := shared.GetAuthenticatedAccount(ctx); account.ID != round.Venture.Business.OwnerAccountID {
+		h.logger.Error("business owner account id does not match authenticated account id",
+			zap.Any("business owner account id", round.Venture.Business.OwnerAccountID),
+			zap.Any("authenticated account id", account.ID))
+
+		return nil, huma.Error403Forbidden("Cannot get investments for a round for a business you do not own")
+	}
+
+	investments, total, err := h.service.RoundService.GetInvestmentsByPage(ctx, input.ID, input.PageSize, input.Page, input.InvestmentFilter)
 
 	if err != nil {
 		switch {
@@ -125,7 +209,7 @@ func (h *httpHandler) getInvestmentsByPage(ctx context.Context, input *shared.Ge
 	resp := &shared.GetOffsetPaginatedRoundInvestmentsOutput{}
 	resp.Body.Message = "Investments fetched successfully"
 	resp.Body.Investments = investments
-
+	resp.Body.Total = total
 	if len(investments) > input.PageSize {
 		resp.Body.HasMore = true
 		resp.Body.Investments = resp.Body.Investments[:len(resp.Body.Investments)-1]
@@ -134,23 +218,15 @@ func (h *httpHandler) getInvestmentsByPage(ctx context.Context, input *shared.Ge
 	return resp, nil
 }
 
-func (h *httpHandler) acceptInvestment(ctx context.Context, input *shared.ParentInvestmentIDParam) (*shared.MessageResponse, error) {
-	err := h.service.RoundService.AcceptInvestment(ctx, input.ID, input.InvestmentID)
-	if err != nil {
-		h.logger.Error("failed to accept investment", zap.Error(err))
-		return nil, huma.Error500InternalServerError("An error occurred while accepting the investment")
-	}
-
-	resp := &shared.MessageResponse{}
-	resp.Message = "Investment accepted successfully"
-
-	return resp, nil
+type GetByCursorInput struct {
+	shared.CursorPaginationRequest
+	round.RoundFilter
 }
 
-func (h *httpHandler) getByCursor(ctx context.Context, input *shared.CursorPaginationRequest) (*shared.GetCursorPaginatedRoundsOutput, error) {
+func (h *httpHandler) getByCursor(ctx context.Context, input *GetByCursorInput) (*shared.GetCursorPaginatedRoundsOutput, error) {
 	limit := input.Limit + 1
 
-	rounds, err := h.service.RoundService.GetByCursor(ctx, limit, input.Cursor)
+	rounds, err := h.service.RoundService.GetByCursor(ctx, limit, input.Cursor, input.RoundFilter)
 
 	if err != nil {
 		switch {
@@ -175,8 +251,13 @@ func (h *httpHandler) getByCursor(ctx context.Context, input *shared.CursorPagin
 	return resp, nil
 }
 
-func (h *httpHandler) getByPage(ctx context.Context, input *shared.OffsetPaginationRequest) (*shared.GetOffsetPaginatedRoundsOutput, error) {
-	rounds, err := h.service.RoundService.GetByPage(ctx, input.PageSize, input.Page)
+type GetByPageInput struct {
+	shared.OffsetPaginationRequest
+	round.RoundFilter
+}
+
+func (h *httpHandler) getByPage(ctx context.Context, input *GetByPageInput) (*shared.GetOffsetPaginatedRoundsOutput, error) {
+	rounds, total, err := h.service.RoundService.GetByPage(ctx, input.PageSize, input.Page, input.RoundFilter)
 
 	if err != nil {
 		switch {
@@ -191,6 +272,7 @@ func (h *httpHandler) getByPage(ctx context.Context, input *shared.OffsetPaginat
 	resp := &shared.GetOffsetPaginatedRoundsOutput{}
 	resp.Body.Message = "Rounds fetched successfully"
 	resp.Body.Rounds = rounds
+	resp.Body.Total = total
 
 	if len(rounds) > input.PageSize {
 		resp.Body.HasMore = true
@@ -225,7 +307,36 @@ type RoundLikeInput struct {
 }
 
 func (h *httpHandler) createLike(ctx context.Context, input *RoundLikeInput) (*shared.MessageOutput, error) {
-	_, err := h.service.RoundService.CreateLike(ctx, round.CreateRoundLikeParams{
+	account := shared.GetAuthenticatedAccount(ctx)
+
+	if account.ID != input.AccountID {
+		h.logger.Error("account id does not match authenticated account id",
+			zap.Any("authenticated account id", account.ID),
+			zap.Any("input account id", input.AccountID))
+
+		return nil, huma.Error403Forbidden("Cannot like for another account")
+	}
+
+	roundRecord, err := h.service.RoundService.GetById(ctx, input.ID)
+	if err != nil {
+		switch {
+		case errors.Is(err, sql.ErrNoRows):
+			return nil, huma.Error404NotFound("Round not found")
+		default:
+			h.logger.Error("failed to fetch round", zap.Error(err))
+			return nil, huma.Error500InternalServerError("An error occurred while fetching the round")
+		}
+	}
+
+	if account.ID != roundRecord.Venture.Business.OwnerAccountID {
+		h.logger.Error("business owner account id does not match authenticated account id",
+			zap.Any("business owner account id", roundRecord.Venture.Business.OwnerAccountID),
+			zap.Any("authenticated account id", account.ID))
+
+		return nil, huma.Error403Forbidden("Cannot like a round for a business you do not own")
+	}
+
+	_, err = h.service.RoundService.CreateLike(ctx, round.CreateRoundLikeParams{
 		RoundID:   input.ID,
 		AccountID: input.AccountID,
 	})
@@ -242,7 +353,36 @@ func (h *httpHandler) createLike(ctx context.Context, input *RoundLikeInput) (*s
 }
 
 func (h *httpHandler) deleteLike(ctx context.Context, input *RoundLikeInput) (*shared.MessageOutput, error) {
-	err := h.service.RoundService.DeleteLike(ctx, input.ID, input.AccountID)
+	account := shared.GetAuthenticatedAccount(ctx)
+
+	if account.ID != input.AccountID {
+		h.logger.Error("account id does not match authenticated account id",
+			zap.Any("authenticated account id", account.ID),
+			zap.Any("input account id", input.AccountID))
+
+		return nil, huma.Error403Forbidden("Cannot delete like for another account")
+	}
+
+	roundRecord, err := h.service.RoundService.GetById(ctx, input.ID)
+	if err != nil {
+		switch {
+		case errors.Is(err, sql.ErrNoRows):
+			return nil, huma.Error404NotFound("Round not found")
+		default:
+			h.logger.Error("failed to fetch round", zap.Error(err))
+			return nil, huma.Error500InternalServerError("An error occurred while fetching the round")
+		}
+	}
+
+	if account.ID != roundRecord.Venture.Business.OwnerAccountID {
+		h.logger.Error("business owner account id does not match authenticated account id",
+			zap.Any("business owner account id", roundRecord.Venture.Business.OwnerAccountID),
+			zap.Any("authenticated account id", account.ID))
+
+		return nil, huma.Error403Forbidden("Cannot delete like for a round for a business you do not own")
+	}
+
+	err = h.service.RoundService.DeleteLike(ctx, input.ID, input.AccountID)
 	if err != nil {
 		h.logger.Error("failed to delete round like", zap.Error(err))
 		return nil, huma.Error500InternalServerError("An error occurred while deleting the round like")
@@ -254,7 +394,18 @@ func (h *httpHandler) deleteLike(ctx context.Context, input *RoundLikeInput) (*s
 	return resp, nil
 }
 
+// TODO: move to account handler
 func (h *httpHandler) isLikedByAccount(ctx context.Context, input *RoundLikeInput) (*shared.IsLikedOutput, error) {
+	account := shared.GetAuthenticatedAccount(ctx)
+
+	if account.ID != input.AccountID {
+		h.logger.Error("account id does not match authenticated account id",
+			zap.Any("authenticated account id", account.ID),
+			zap.Any("input account id", input.AccountID))
+
+		return nil, huma.Error403Forbidden("Cannot check if round is liked by another account")
+	}
+
 	liked, err := h.service.RoundService.IsLikedByAccount(ctx, input.ID, input.AccountID)
 	if err != nil {
 		h.logger.Error("failed to check if round is liked by account", zap.Error(err))
